+++ terra-jupyter-base/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.1.1 - 2023-07-06
+
+- Change notebook config to server config
+
 ## 1.1.0 - 2023-06-23
 
 - Upgrade python 3.7 to 3.10