+++ terra-jupyter-hail/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.1.1 - 2023-07-06
+
+- Change notebook config to server config
+
+Image URL: `us.gcr.io/broad-dsp-gcr-public/terra-jupyter-hail:1.1.1`
+
 ## 1.1.0 - 2023-06-23
 
 - Update python 3.7 to 3.10